refacotr: Improve container, view structure
