+++ src/features/game1/container/index.js
@@ -1,0 +1,4 @@
+import FaceRecognitionContainer from "./FaceRecognitionContainer";
+import Game1Container from "./Game1Container";
+
+export { FaceRecognitionContainer, Game1Container };
