+++ src/features/game1/presenter/index.js
@@ -1,0 +1,6 @@
+import FaceRecognition from "../presenter/FaceRecognition";
+import Game1 from "../presenter/Game1";
+import Map from "../presenter/Map";
+FaceRecognition;
+
+export { Game1, Map, FaceRecognition };
